+++ packages/pluggableWidgets/datagrid-text-filter-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+-   We resolved an issue where the default value was not working in certain cases.
+
 ## [2.5.0] - 2024-03-27
 
 ### Added